More messing around with CORs issues
Also included a nice little tool to start
the frontend and the backend at the same time.
I wish I was able to see the backend logs though
so I may need to find a way to split the terminal
in half so the backend does not run in the background.